add more methods exist in original combined json parser
